Wad lumps should be addressed by int, not by short
* New SString method: noCaseStartsWith
* Purged most use of short, except for certain data and rendering stuff
* Deleted some redundant std vector init
* Use SString methods where appropriate

+++ src/bsp.h
@@ -109,7 +109,7 @@ typedef enum
 build_result_e;
 
 
-build_result_e AJBSP_BuildLevel(nodebuildinfo_t *info, short lev_idx);
+build_result_e AJBSP_BuildLevel(nodebuildinfo_t *info, int lev_idx);
 
 
 //======================================================================
